+++ app/ui/dialogs/edit_asset_dialog.py
@@ -1,0 +1,4 @@
+from .add_asset_dialog import AddAssetDialog
+
+class EditAssetDialog(AddAssetDialog):
+    pass